+++ 2021_fp/src/main/java/net/koreate/cboard/vo/CommunityCommentVO.java
@@ -16,5 +16,5 @@ public class CommunityCommentVO {
 	private String ccshowboard; // 추가
 	private int uno;
 	
-	private int position;
+	private double position;
 }
